requirements.txt
streamlit
torch
transformers
boto3
python-dotenv
